Test.HUnit.DejaFu

Description

This module allows using Deja Fu predicates with HUnit to test the behaviour of concurrent systems.

Testing

testAuto

Arguments

:: (Eq a, Show a)

=> (forall t. Conc t a)
The computation to test

-> ConcTest a

Automatically test a computation. In particular, look for deadlocks, uncaught exceptions, and multiple return values.

This uses the Conc monad for testing, which is an instance of MonadConc. If you need to test something which also uses MonadIO, use testAutoIO.

testDejafu

Arguments

:: (Eq a, Show a)

=> (forall t. Conc t a)
The computation to test

-> Predicate a
The predicate to check

-> ConcTest a

Check that a predicate holds.

testDejafus

Arguments

:: (Eq a, Show a)

=> (forall t. Conc t a)
The computation to test

-> [Predicate a]
The list of predicates to check

-> ConcTest a

Variant of testDejafu which takes a collection of predicates to test. This will share work between the predicates, rather than running the concurrent computation many times for each predicate.

testAutoIO :: (Eq a, Show a) => (forall t. ConcIO t a) -> ConcIOTest a

Variant of testAuto for computations which do IO.

testDejafuIO :: (Eq a, Show a) => (forall t. ConcIO t a) -> Predicate a -> ConcIOTest a

Variant of testDejafu for computations which do IO.

testDejafusIO :: (Eq a, Show a) => (forall t. ConcIO t a) -> [Predicate a] -> ConcIOTest a

Variant of testDejafus for computations which do IO.

Testing under Relaxed Memory

data MemType :: *

Constructors

SequentialConsistency

TotalStoreOrder

PartialStoreOrder

testAuto'

Arguments

:: (Eq a, Show a)

=> MemType
The memory model to use for non-synchronised CRef operations.

-> (forall t. Conc t a)
The computation to test

-> ConcTest a

Variant of testAuto which tests a computation under a given memory model.

testAutoIO' :: (Eq a, Show a) => MemType -> (forall t. ConcIO t a) -> ConcIOTest a

Variant of testAuto' for computations which do IO.

testDejafus'

Arguments

:: (Eq a, Show a)

=> MemType
The memory model to use for non-synchronised CRef operations.

-> Int
The maximum number of pre-emptions to allow in a single execution

-> (forall t. Conc t a)
The computation to test

-> [Predicate a]
The list of predicates to check

-> ConcTest a

Variant of testDejafus which takes a memory model and pre-emption bound.

testDejafusIO' :: (Eq a, Show a) => MemType -> Int -> (forall t. ConcIO t a) -> [Predicate a] -> ConcIOTest a

Variant of dejafus' for computations which do IO.